requirements.txt
asgiref==3.7.2
branca==0.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
decouple==0.0.7
Django==4.2.2
django-cors-headers==4.2.0
django-filter==23.2
django-geoposition==0.3.0
django-geoposition-2==0.4.0
django-google-maps==0.13.0
django-utils==0.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.3
geocoder==1.38.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
Pillow==9.5.0
PyJWT==2.8.0
pypng==0.20220715.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
qrcode==7.4.2
ratelim==0.1.6
reportlab==4.0.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
